Reformat TomcatOperator & add to pom structure
