+++ pom.xml
@@ -78,7 +78,8 @@
         <module>operator-framework-junit5</module>
         <module>operator-framework</module>
         <module>samples</module>
-      <module>micrometer-support</module>
+        <module>micrometer-support</module>
+        <module>sample-operators</module>
     </modules>
 
 
@@ -334,6 +335,7 @@
                             <includes>
                                 <include>**/*Test.java</include>
                                 <include>**/*IT.java</include>
+                                <include>**/*E2E.java</include>
                             </includes>
                         </configuration>
                     </plugin>
@@ -353,6 +355,27 @@
                             </includes>
                             <excludes>
                                 <exclude>**/*Test.java</exclude>
+                                <exclude>**/*E2E.java</exclude>
+                            </excludes>
+                        </configuration>
+                    </plugin>
+                </plugins>
+            </build>
+        </profile>
+        <profile>
+            <id>end-to-end-tests</id>
+            <build>
+                <plugins>
+                    <plugin>
+                        <groupId>org.apache.maven.plugins</groupId>
+                        <artifactId>maven-surefire-plugin</artifactId>
+                        <configuration>
+                            <includes>
+                                <include>**/*E2E.java</include>
+                            </includes>
+                            <excludes>
+                                <exclude>**/*Test.java</exclude>
+                                <exclude>**/*IT.java</exclude>
                             </excludes>
                         </configuration>
                     </plugin>
