+++ sample-operators/pom.xml
@@ -6,21 +6,22 @@
 
     <parent>
         <groupId>io.javaoperatorsdk</groupId>
-        <artifactId>sample-operators</artifactId>
+        <artifactId>java-operator-sdk</artifactId>
         <version>1.9.9-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
 
-    <artifactId>sample-tomcat-operator</artifactId>
-    <name>Operator SDK - Samples - Tomcat</name>
-    <description>Provisions Tomcat Pods and deploys Webapplications in them</description>
-    <packaging>jar</packaging>
+    <artifactId>sample-operators</artifactId>
+    <name>Operator SDK - Samples</name>
+    <packaging>pom</packaging>
 
     <properties>
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
         <jib-maven-plugin.version>3.1.4</jib-maven-plugin.version>
     </properties>
+
+    <modules>
+        <module>tomcat-operator</module>
+    </modules>
 
     <dependencies>
         <dependency>
