+++ sample-operators/tomcat-operator/src/main/java/io/javaoperatorsdk/operator/sample/WebappController.java
@@ -1,5 +1,16 @@
 package io.javaoperatorsdk.operator.sample;
 
+import java.io.ByteArrayOutputStream;
+import java.util.List;
+import java.util.Objects;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import io.fabric8.kubernetes.api.model.Pod;
 import io.fabric8.kubernetes.api.model.apps.Deployment;
 import io.fabric8.kubernetes.client.KubernetesClient;
@@ -7,19 +18,8 @@ import io.fabric8.kubernetes.client.dsl.ExecListener;
 import io.fabric8.kubernetes.client.dsl.ExecWatch;
 import io.javaoperatorsdk.operator.api.*;
 import io.javaoperatorsdk.operator.processing.event.EventSourceManager;
-import okhttp3.Response;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import java.io.ByteArrayOutputStream;
-import java.io.InputStream;
-import java.nio.charset.Charset;
-import java.util.List;
-import java.util.Objects;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
+import okhttp3.Response;
 
 @Controller
 public class WebappController implements ResourceController<Webapp> {
@@ -34,30 +34,41 @@ public class WebappController implements ResourceController<Webapp> {
 
   @Override
   public void init(EventSourceManager eventSourceManager) {
-    TomcatEventSource tomcatEventSource = TomcatEventSource.createAndRegisterWatch(kubernetesClient);
+    TomcatEventSource tomcatEventSource =
+        TomcatEventSource.createAndRegisterWatch(kubernetesClient);
     eventSourceManager.registerEventSource("tomcat-event-source", tomcatEventSource);
   }
 
   /**
-   * This method will be called not only on changes to Webapp objects but also when Tomcat objects change.
+   * This method will be called not only on changes to Webapp objects but also when Tomcat objects
+   * change.
    */
   @Override
   public UpdateControl<Webapp> createOrUpdateResource(Webapp webapp, Context<Webapp> context) {
-    if (webapp.getStatus() != null && Objects.equals(webapp.getSpec().getUrl(), webapp.getStatus().getDeployedArtifact())) {
+    if (webapp.getStatus() != null
+        && Objects.equals(webapp.getSpec().getUrl(), webapp.getStatus().getDeployedArtifact())) {
       return UpdateControl.noUpdate();
     }
 
     var tomcatClient = kubernetesClient.customResources(Tomcat.class);
-    Tomcat tomcat = tomcatClient.inNamespace(webapp.getMetadata().getNamespace()).withName(webapp.getSpec().getTomcat()).get();
+    Tomcat tomcat = tomcatClient.inNamespace(webapp.getMetadata().getNamespace())
+        .withName(webapp.getSpec().getTomcat()).get();
     if (tomcat == null) {
-      throw new IllegalStateException("Cannot find Tomcat " + webapp.getSpec().getTomcat() + " for Webapp " + webapp.getMetadata().getName() + " in namespace " + webapp.getMetadata().getNamespace());
+      throw new IllegalStateException("Cannot find Tomcat " + webapp.getSpec().getTomcat()
+          + " for Webapp " + webapp.getMetadata().getName() + " in namespace "
+          + webapp.getMetadata().getNamespace());
     }
 
-    if (tomcat.getStatus() != null && Objects.equals(tomcat.getSpec().getReplicas(), tomcat.getStatus().getReadyReplicas())) {
-      log.info("Tomcat is ready and webapps not yet deployed. Commencing deployment of {} in Tomcat {}", webapp.getMetadata().getName(), tomcat.getMetadata().getName());
-      String[] command = new String[]{"wget", "-O", "/data/" + webapp.getSpec().getContextPath() + ".war", webapp.getSpec().getUrl()};
-      if(log.isInfoEnabled()){
-        command = new String[]{"time", "wget", "-O", "/data/" + webapp.getSpec().getContextPath() + ".war", webapp.getSpec().getUrl()};
+    if (tomcat.getStatus() != null
+        && Objects.equals(tomcat.getSpec().getReplicas(), tomcat.getStatus().getReadyReplicas())) {
+      log.info(
+          "Tomcat is ready and webapps not yet deployed. Commencing deployment of {} in Tomcat {}",
+          webapp.getMetadata().getName(), tomcat.getMetadata().getName());
+      String[] command = new String[] {"wget", "-O",
+          "/data/" + webapp.getSpec().getContextPath() + ".war", webapp.getSpec().getUrl()};
+      if (log.isInfoEnabled()) {
+        command = new String[] {"time", "wget", "-O",
+            "/data/" + webapp.getSpec().getContextPath() + ".war", webapp.getSpec().getUrl()};
       }
 
       String[] commandStatusInAllPods = executeCommandInAllPods(kubernetesClient, webapp, command);
@@ -70,7 +81,8 @@ public class WebappController implements ResourceController<Webapp> {
       return UpdateControl.updateStatusSubResource(webapp);
     } else {
       log.info("WebappController invoked but Tomcat not ready yet ({}/{})",
-              tomcat.getStatus() != null ? tomcat.getStatus().getReadyReplicas() : 0, tomcat.getSpec().getReplicas());
+          tomcat.getStatus() != null ? tomcat.getStatus().getReadyReplicas() : 0,
+          tomcat.getSpec().getReplicas());
       return UpdateControl.noUpdate();
     }
   }
@@ -108,7 +120,7 @@ public class WebappController implements ResourceController<Webapp> {
               .list()
               .getItems();
       status = new String[pods.size()];
-      for (int i=0; i<pods.size();i++) {
+      for (int i = 0; i < pods.size(); i++) {
         Pod pod = pods.get(i);
         log.info(
             "Executing command {} in Pod {}",
@@ -117,13 +129,14 @@ public class WebappController implements ResourceController<Webapp> {
 
         CompletableFuture<String> data = new CompletableFuture<>();
         try (ExecWatch execWatch = execCmd(pod, data, command)) {
-          status[i] = ""+pod.getMetadata().getName()+":"+data.get(30, TimeUnit.SECONDS);;
+          status[i] = "" + pod.getMetadata().getName() + ":" + data.get(30, TimeUnit.SECONDS);;
         } catch (ExecutionException e) {
-          status[i] = ""+pod.getMetadata().getName()+": ExecutionException - "+e.getMessage();
+          status[i] = "" + pod.getMetadata().getName() + ": ExecutionException - " + e.getMessage();
         } catch (InterruptedException e) {
-          status[i] = ""+pod.getMetadata().getName()+": InterruptedException - "+e.getMessage();
+          status[i] =
+              "" + pod.getMetadata().getName() + ": InterruptedException - " + e.getMessage();
         } catch (TimeoutException e) {
-          status[i] = ""+pod.getMetadata().getName()+": TimeoutException - "+e.getMessage();
+          status[i] = "" + pod.getMetadata().getName() + ": TimeoutException - " + e.getMessage();
         }
       }
     }
@@ -133,13 +146,13 @@ public class WebappController implements ResourceController<Webapp> {
   private ExecWatch execCmd(Pod pod, CompletableFuture<String> data, String... command) {
     ByteArrayOutputStream baos = new ByteArrayOutputStream();
     return kubernetesClient.pods()
-            .inNamespace(pod.getMetadata().getNamespace())
-            .withName(pod.getMetadata().getName())
-            .inContainer("war-downloader")
-            .writingOutput(baos)
-            .writingError(baos)
-            .usingListener(new SimpleListener(data, baos))
-            .exec(command);
+        .inNamespace(pod.getMetadata().getNamespace())
+        .withName(pod.getMetadata().getName())
+        .inContainer("war-downloader")
+        .writingOutput(baos)
+        .writingError(baos)
+        .usingListener(new SimpleListener(data, baos))
+        .exec(command);
   }
 
   static class SimpleListener implements ExecListener {
@@ -147,6 +160,7 @@ public class WebappController implements ResourceController<Webapp> {
     private CompletableFuture<String> data;
     private ByteArrayOutputStream baos;
     private final Logger log = LoggerFactory.getLogger(getClass());
+
     public SimpleListener(CompletableFuture<String> data, ByteArrayOutputStream baos) {
       this.data = data;
       this.baos = baos;
